feat(gui): separate admin & other change methods
The idea of the comments in the Rust is that they will eventually be
generated by a Rust macro that replaces all the `assert_owner` calls.
The macro can look like:

    #[only(::owner, willem.near, ::admins, ::get_admin_contract_id::admins)]

Or maybe:

    #[only(owner, "willem.near", admins, ::get_admin_contract_id::admins)]

The "owner" view method could be auto-generated from a `derive` on the
struct:

    #[derive(BorshDeserialize, BorshSerialize, PanicOnDefault, Ownable)]

While `::admins` and `::get_admin_contract_id` would be
explicitly-defined methods on this contract.

The methods specified here would all be expected to return `AccountId |
Vec<AccountId>`

::get_admin_contract_id::admins specifies that any accounts returned by
this method on an external contract should be allowed calling the
specified method. This would be simple to handle on a frontend, but
implies that a cross-contract call with callback is generated by having
such a specification in the macro.