feat(gui): remove assumptions about admins methods
This removes entanglement between what will soon become a
general-purpose frontend and the admins behavior of the TenK contract.
Instead, it only assumes that a specific comment format is followed in
the contract. In the future these comments will be generated in the wit
if an admins macro is present for the given contract method.

Now `Change Methods` in the UI will be longer or shorter, depending on
which methods the signed-in user is able to call.